HT okHttp GET and POST request on twetter(u1 - date format)

+++ app/src/main/java/by/twitter/model/Tweet.kt
@@ -15,7 +15,7 @@ data class Tweet(
         @SerializedName("user")
         val user: User,
         @SerializedName("retweet_count")
-        val retweetCount: Long,
+        val retweetCount: Int,
         @SerializedName("favorite_count")
-        val favoriteCount: Long
+        val favoriteCount: Int
 ) : Parcelable
